nu-project/image-agent/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
replicate>=0.25.0
supabase>=2.0.0
cryptography>=42.0.0
requests>=2.31.0
